+++ random_produce.h
@@ -5,6 +5,8 @@
 #include <stdlib.h>
 #include <time.h>
 #include <math.h>
+#include <inttypes.h>
 
-long double H_function(long double y[], char normalize, int Y);
+//long double H_function(long double y[], char normalize, uint32_t Y);
+double H_function(double y[], char normalize, uint32_t Y);
 #endif
